[PVR] Refactor CPVRGUIActions step 8: Move database related actions to own class.
