+++ xbmc/pvr/guilib/PVRGUIActions.h
@@ -9,6 +9,7 @@
 #pragma once
 
 #include "pvr/guilib/PVRGUIActionsChannels.h"
+#include "pvr/guilib/PVRGUIActionsDatabase.h"
 #include "pvr/guilib/PVRGUIActionsEPG.h"
 #include "pvr/guilib/PVRGUIActionsParentalControl.h"
 #include "pvr/guilib/PVRGUIActionsPlayback.h"
@@ -26,6 +27,7 @@ class CFileItem;
 namespace PVR
 {
 class CPVRGUIActions : public CPVRGUIActionsChannels,
+                       public CPVRGUIActionsDatabase,
                        public CPVRGUIActionsEPG,
                        public CPVRGUIActionsParentalControl,
                        public CPVRGUIActionsPlayback,
@@ -42,13 +44,6 @@ public:
    * @return true on success, false otherwise.
    */
   bool ProcessSettingsMenuHooks();
-
-  /*!
-   * @brief Reset the TV database to it's initial state and delete all the data.
-   * @param bResetEPGOnly True to only reset the EPG database, false to reset both PVR and EPG database.
-   * @return true on success, false otherwise.
-   */
-  bool ResetPVRDatabase(bool bResetEPGOnly);
 
   /*!
    * @brief Get the currently selected item path; used across several windows/dialogs to share item selection.
